+++ RestaurantReservation.Db/Models/MenuItem.cs
@@ -2,8 +2,8 @@
 
 public class MenuItem
 {
-    public int MenuItemId { get; set; }
-    public int RestaurantId { get; set; }
+    public long MenuItemId { get; set; }
+    public long RestaurantId { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
     public decimal Price { get; set; }
