+++ RestaurantReservation.Db/Models/OrderItem.cs
@@ -2,9 +2,9 @@
 
 public class OrderItem
 {
-    public int OrderItemId { get; set; }
-    public int OrderId { get; set; }
-    public int MenuItemId { get; set; }
+    public long OrderItemId { get; set; }
+    public long OrderId { get; set; }
+    public long MenuItemId { get; set; }
     public int Quantity { get; set; }
 
     public Order Order { get; set; }
