+++ RestaurantReservation.Db/Models/Restaurant.cs
@@ -2,7 +2,7 @@
 
 public class Restaurant
 {
-    public int RestaurantId { get; set; }
+    public long RestaurantId { get; set; }
     public string Name { get; set; }
     public string Address { get; set; }
     public string PhoneNumber { get; set; }
