+++ RestaurantReservation.Db/Models/Table.cs
@@ -2,9 +2,9 @@
 
 public class Table
 {
-    public int TableId { get; set; }
-    public int RestaurantId { get; set; }
-    public int ReservationId { get; set; }
+    public long TableId { get; set; }
+    public long RestaurantId { get; set; }
+    public long ReservationId { get; set; }
     public int Capacity { get; set; }
 
     public Reservation Reservation { get; set; }
